refactor: split enforcer and user policy
