+++ internal/authorizer/enforcer_test.go
@@ -86,19 +86,20 @@ func Test_IAuthorizer_AddPolicy(t *testing.T) {
 	assert.Equal(t, false, allow)
 }
 
-func Test_IAuthorizer_LoginAndSession(t *testing.T) {
-	a, err := authorizer.New(_conf)
-	require.Nil(t, err)
-
-	_, tokenString, err := a.Login("root", "123456")
-	require.Nil(t, err)
-	t.Log(tokenString)
-
-	token, err := authorizer.Session(tokenString)
-	require.Nil(t, err)
-
-	assert.NotEmpty(t, token.UserId)
-}
+//
+//func Test_IAuthorizer_LoginAndSession(t *testing.T) {
+//	a, err := authorizer.New(_conf)
+//	require.Nil(t, err)
+//
+//	_, tokenString, err := a.Login("root", "123456")
+//	require.Nil(t, err)
+//	t.Log(tokenString)
+//
+//	token, err := authorizer.Session(tokenString)
+//	require.Nil(t, err)
+//
+//	assert.NotEmpty(t, token.UserId)
+//}
 
 func Test_IAuthorizer_ListPolicy(t *testing.T) {
 	a, err := authorizer.New(_conf)
